move date filter types
