+++ src/components/date-filter.tsx
@@ -1,3 +1,16 @@
+export interface DateRange {
+  startDate: Date | null;
+  endDate: Date | null;
+  startTime?: string; // HH:mm format
+  endTime?: string; // HH:mm format
+}
+
+export interface DateFilterPreset {
+  id: string;
+  label: string;
+  getDateRange: () => DateRange;
+}
+
 import {
   Alert,
   Button,
@@ -10,7 +23,6 @@ import {
 import { LuCalendar, LuX } from "react-icons/lu";
 import useStacMap from "../hooks/stac-map";
 import { DATE_FILTER_PRESETS, isValidDateRange } from "../utils/date-filter";
-import type { DateRange } from "../types/stac";
 
 export default function DateFilter() {
   const { dateRange, setDateRange, clearDateRange, isDateFilterActive } =
