+++ src/types/stac.ts
@@ -35,19 +35,7 @@ export interface StacSearchRequest {
   link: StacLink;
 }
 
-// Add new date range types
-export interface DateRange {
-  startDate: Date | null;
-  endDate: Date | null;
-  startTime?: string; // HH:mm format
-  endTime?: string; // HH:mm format
-}
 
-export interface DateFilterPreset {
-  id: string;
-  label: string;
-  getDateRange: () => DateRange;
-}
 
 export interface StacGeoparquetMetadata {
   count: number;
